backend/requirements.txt
fastapi[all]==0.111.0
sqlalchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
uvicorn[standard]==0.29.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
